requirements.txt
anthropic>=0.25.0
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.26.0
matplotlib>=3.8.0
requests>=2.31.0
python-dotenv>=1.0.0
weasyprint>=60.0    # optional, for PDF
